Move apispec & fake data to test_utils module
