+++ mithril-aggregator/src/http_server/routes/certificate_routes.rs
@@ -82,9 +82,9 @@ mod tests {
     const API_SPEC_FILE: &str = "../openapi.yaml";
 
     use crate::http_server::SERVER_BASE_PATH;
-    use mithril_common::apispec::APISpec;
     use mithril_common::store::adapter::FailStoreAdapter;
-    use mithril_common::{entities, fake_data};
+    use mithril_common::test_utils::apispec::APISpec;
+    use mithril_common::{entities, test_utils::fake_data};
     use serde_json::Value::Null;
     use warp::http::Method;
     use warp::test::request;
