+++ mithril-aggregator/src/http_server/routes/epoch_routes.rs
@@ -84,7 +84,7 @@ mod tests {
     const API_SPEC_FILE: &str = "../openapi.yaml";
 
     use crate::http_server::SERVER_BASE_PATH;
-    use mithril_common::apispec::APISpec;
+    use mithril_common::test_utils::apispec::APISpec;
     use serde_json::Value::Null;
     use warp::http::Method;
     use warp::test::request;
