+++ mithril-aggregator/src/http_server/routes/signatures_routes.rs
@@ -56,8 +56,8 @@ mod tests {
     const API_SPEC_FILE: &str = "../openapi.yaml";
 
     use crate::http_server::SERVER_BASE_PATH;
-    use mithril_common::apispec::APISpec;
-    use mithril_common::fake_data;
+    use mithril_common::test_utils::apispec::APISpec;
+    use mithril_common::test_utils::fake_data;
     use tokio::sync::RwLock;
     use warp::http::Method;
     use warp::test::request;
