+++ mithril-common/src/lib.rs
@@ -10,8 +10,6 @@
 //! - A [certificate chain] used to validate the Certificate Chain created by an aggregator
 //! - The [entities] used by, and exchanged between, the aggregator, signers and client.
 
-#[cfg(any(test, feature = "test_only"))]
-pub mod apispec;
 mod beacon_provider;
 pub mod certificate_chain;
 pub mod chain_observer;
@@ -19,8 +17,6 @@ pub mod crypto_helper;
 pub mod database;
 pub mod digesters;
 pub mod entities;
-#[cfg(any(test, feature = "test_only"))]
-pub mod fake_data;
 pub mod sqlite;
 pub mod store;
 pub mod test_utils;
